+++ test/unit/spec_helper.rb
@@ -1,0 +1,5 @@
+require 'chefspec'
+ChefSpec::Coverage.start!
+require 'chefspec/berkshelf'
+
+Dir['./test/unit/support/*.rb'].sort.each { |f| require f }